+++ sql/core/src/test/resources/sql-tests/inputs/subquery/exists-subquery/exists-having.sql
@@ -1,36 +1,33 @@
 -- Tests HAVING clause in subquery.
 
-CREATE TEMPORARY VIEW EMP AS SELECT * FROM VALUES
-  (100, "emp 1", date "2005-01-01", 100.00D, 10),
-  (100, "emp 1", date "2005-01-01", 100.00D, 10),
-  (200, "emp 2", date "2003-01-01", 200.00D, 10),
-  (300, "emp 3", date "2002-01-01", 300.00D, 20),
-  (400, "emp 4", date "2005-01-01", 400.00D, 30),
-  (500, "emp 5", date "2001-01-01", 400.00D, NULL),
-  (600, "emp 6 - no dept", date "2001-01-01", 400.00D, 100),
-  (700, "emp 7", date "2010-01-01", 400.00D, 100),
-  (800, "emp 8", date "2016-01-01", 150.00D, 70)
-AS EMP(id, emp_name, hiredate, salary, dept_id);
+CREATE TEMPORARY VIEW EMP(id, emp_name, hiredate, salary, dept_id) AS VALUES
+  (100, 'emp 1', date '2005-01-01', double(100.00), 10),
+  (100, 'emp 1', date '2005-01-01', double(100.00), 10),
+  (200, 'emp 2', date '2003-01-01', double(200.00), 10),
+  (300, 'emp 3', date '2002-01-01', double(300.00), 20),
+  (400, 'emp 4', date '2005-01-01', double(400.00), 30),
+  (500, 'emp 5', date '2001-01-01', double(400.00), NULL),
+  (600, 'emp 6 - no dept', date '2001-01-01', double(400.00), 100),
+  (700, 'emp 7', date '2010-01-01', double(400.00), 100),
+  (800, 'emp 8', date '2016-01-01', double(150.00), 70);
 
-CREATE TEMPORARY VIEW DEPT AS SELECT * FROM VALUES
-  (10, "dept 1", "CA"),
-  (20, "dept 2", "NY"),
-  (30, "dept 3", "TX"),
-  (40, "dept 4 - unassigned", "OR"),
-  (50, "dept 5 - unassigned", "NJ"),
-  (70, "dept 7", "FL")
-AS DEPT(dept_id, dept_name, state);
+CREATE TEMPORARY VIEW DEPT(dept_id, dept_name, state) AS VALUES
+  (10, 'dept 1', 'CA'),
+  (20, 'dept 2', 'NY'),
+  (30, 'dept 3', 'TX'),
+  (40, 'dept 4 - unassigned', 'OR'),
+  (50, 'dept 5 - unassigned', 'NJ'),
+  (70, 'dept 7', 'FL');
 
-CREATE TEMPORARY VIEW BONUS AS SELECT * FROM VALUES
-  ("emp 1", 10.00D),
-  ("emp 1", 20.00D),
-  ("emp 2", 300.00D),
-  ("emp 2", 100.00D),
-  ("emp 3", 300.00D),
-  ("emp 4", 100.00D),
-  ("emp 5", 1000.00D),
-  ("emp 6 - no dept", 500.00D)
-AS BONUS(emp_name, bonus_amt);
+CREATE TEMPORARY VIEW BONUS(emp_name, bonus_amt) AS VALUES
+  ('emp 1', double(10.00)),
+  ('emp 1', double(20.00)),
+  ('emp 2', double(300.00)),
+  ('emp 2', double(100.00)),
+  ('emp 3', double(300.00)),
+  ('emp 4', double(100.00)),
+  ('emp 5', double(1000.00)),
+  ('emp 6 - no dept', double(500.00));
 
 -- simple having in subquery. 
 -- TC.01.01
