+++ sql/core/src/test/resources/sql-tests/inputs/subquery/in-subquery/in-count-bug.sql
@@ -1,3 +1,4 @@
+--ONLY_IF spark
 create temporary view t1(c1, c2) as values (0, 1), (1, 2);
 create temporary view t2(c1, c2) as values (0, 2), (0, 3);
 create temporary view t3(c1, c2) as values (0, 3), (1, 4), (2, 5);
