+++ sql/core/src/test/resources/sql-tests/inputs/subquery/subquery-in-from.sql
@@ -1,3 +1,4 @@
+--ONLY_IF spark
 -- Aliased subqueries in FROM clause
 SELECT * FROM (SELECT * FROM testData) AS t WHERE key = 1;
